feat: Move AuthenticationMode to PI.py
